backend/image_service/requirements.txt
Flask==1.1.4
flask-cors==3.0.10
requests==2.25.1
markupsafe==2.0.1
Pillow==8.2.0
opencv-python-headless==4.5.3.56
flask-cors